+++ lua/ulib/shared/defines.lua
@@ -477,7 +477,6 @@ ULib.UCL_LOAD_DEFAULT = true -- Set this to false to ignore the SetUserGroup() c
 ULib.UCL_USERS = "data/ulib/users.txt"
 ULib.UCL_GROUPS = "data/ulib/groups.txt"
 ULib.UCL_REGISTERED = "data/ulib/misc_registered.txt" -- Holds access strings that ULib has already registered
-ULib.BANS_FILE = "data/ulib/bans.txt"
 
 ULib.DEFAULT_GRANT_ACCESS = { allow={}, deny={}, guest=true }
 end