+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ULib Changelog
 
 ## v2.64 - *(00/00/00)*
+* [FIX] Lua error with unspecified optional player arguments from server console (thanks, Timmy).
 
 ## v2.63 - *(03/21/17)*
 * [FIX] Server errors when loading massive ban lists.